+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/json/ImageToJsonTranslator.java
@@ -87,6 +87,9 @@ public class ImageToJsonTranslator {
       template.addLayerDiffId(layer.getDiffId());
     }
 
+    // Sets the creation time. Instant#toString() returns an ISO-8601 formatted string.
+    template.setCreated(image.getCreated() == null ? null : image.getCreated().toString());
+
     // Adds the environment variables.
     template.setContainerEnvironment(image.getEnvironment());