+++ jib-core/src/test/java/com/google/cloud/tools/jib/registry/credentials/DockerConfigCredentialRetrieverTest.java
@@ -118,4 +118,47 @@ public class DockerConfigCredentialRetrieverTest {
 
     Assert.assertNull(dockerConfigCredentialRetriever.retrieve());
   }
+
+  @Test
+  public void testRetrieve_credentialFromAlias() throws IOException {
+    Mockito.when(mockDockerCredentialHelperFactory.withCredentialHelperSuffix(Mockito.anyString()))
+        .thenReturn(Mockito.mock(DockerCredentialHelper.class));
+    Mockito.when(
+            mockDockerCredentialHelperFactory.withCredentialHelperSuffix(
+                "index.docker.io credential helper"))
+        .thenReturn(mockDockerCredentialHelper);
+
+    DockerConfigCredentialRetriever dockerConfigCredentialRetriever =
+        new DockerConfigCredentialRetriever(
+            "registry.hub.docker.com", dockerConfigFile, mockDockerCredentialHelperFactory);
+
+    Authorization authorization = dockerConfigCredentialRetriever.retrieve();
+    Assert.assertEquals(mockAuthorization, authorization);
+  }
+
+  @Test
+  public void testRetrieve_suffixMatching() throws IOException, URISyntaxException {
+    Path dockerConfigFile =
+        Paths.get(Resources.getResource("json/dockerconfig_index_docker_io_v1.json").toURI());
+
+    DockerConfigCredentialRetriever dockerConfigCredentialRetriever =
+        new DockerConfigCredentialRetriever(
+            "index.docker.io", dockerConfigFile, mockDockerCredentialHelperFactory);
+
+    Authorization authorization = dockerConfigCredentialRetriever.retrieve();
+    Assert.assertEquals("token for index.docker.io/v1/", authorization.getToken());
+  }
+
+  @Test
+  public void testRetrieve_suffixMatchingFromAlias() throws IOException, URISyntaxException {
+    Path dockerConfigFile =
+        Paths.get(Resources.getResource("json/dockerconfig_index_docker_io_v1.json").toURI());
+
+    DockerConfigCredentialRetriever dockerConfigCredentialRetriever =
+        new DockerConfigCredentialRetriever(
+            "registry.hub.docker.com", dockerConfigFile, mockDockerCredentialHelperFactory);
+
+    Authorization authorization = dockerConfigCredentialRetriever.retrieve();
+    Assert.assertEquals("token for index.docker.io/v1/", authorization.getToken());
+  }
 }